Refactored packages and filenames
