+++ GameData/playerData.csv
@@ -2,3 +2,5 @@ Sindre,1
 Kasper,2
 Luna,3
 Mio,4
+n,2
+p,2
