--- src/main/java/edu/ntnu/idi/idattx2002/Modules/Board/LandAction.java
@@ -1,7 +1,0 @@
-package edu.ntnu.idi.idattx2002.Modules.Board;
-import edu.ntnu.idi.idattx2002.Modules.Player.Player;
-
-public interface LandAction {
-  void perform(Player player);
-}
-
