--- src/main/java/edu/ntnu/idi/idattx2002/Modules/DiceObserver.java
@@ -1,9 +1,0 @@
-package edu.ntnu.idi.idattx2002.Modules;
-
-import edu.ntnu.idi.idattx2002.Modules.Dice.Dice;
-import edu.ntnu.idi.idattx2002.view.DiceWindow;
-
-public interface DiceObserver {
-
-  public void update(Dice dice);
-}
