--- src/main/java/edu/ntnu/idi/idattx2002/Modules/Player/PlayerDAO.java
@@ -1,61 +1,0 @@
-package edu.ntnu.idi.idattx2002.Modules.Player;
-import java.io.*;
-import java.util.ArrayList;
-
-public class PlayerDAO {
-
-    public PlayerDAO() {}
-
-
-    public void writePlayer(String name, int imgID) throws IOException {
-        FileWriter writer = new FileWriter("GameData/playerData.csv", true);
-        writer.write(name +","+imgID+"\n");
-        writer.close();
-    }
-
-    public ArrayList<ArrayList<String>> getPlayers() throws IOException {
-        ArrayList<ArrayList<String>> players = new ArrayList<>();
-
-        BufferedReader reader = new BufferedReader(new FileReader("GameData/playerData.csv"));
-
-        String line;
-        while ((line = reader.readLine()) != null) {
-            ArrayList<String> player = new ArrayList<>();
-            String[] parts = line.split(",");
-            player.add(parts[0]);
-            player.add(parts[1]);
-            players.add(player);
-        }
-
-        reader.close();
-
-        return players;
-    }
-
-    public void deletePlayer(String nameToDelete) throws IOException {
-        ArrayList<String> updatedLines = new ArrayList<>();
-
-        try (BufferedReader reader = new BufferedReader(new FileReader("GameData/playerData.csv"))) {
-            String line;
-
-            while ((line = reader.readLine()) != null) {
-                if (!line.startsWith(nameToDelete + ",")) {
-                    updatedLines.add(line);
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-            return;
-        }
-
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter("GameData/playerData.csv"))) {
-            for (String updatedLine : updatedLines) {
-                writer.write(updatedLine);
-                writer.newLine();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-}
