--- src/main/java/edu/ntnu/idi/idattx2002/Modules/PlayerObserver.java
@@ -1,9 +1,0 @@
-package edu.ntnu.idi.idattx2002.Modules;
-
-import edu.ntnu.idi.idattx2002.Modules.Player.Player;
-
-public interface PlayerObserver {
-
-  void update(Player player);
-
-}
