+++ src/main/java/edu/ntnu/idi/idattx2002/logic/Board/Actions/LadderAction.java
@@ -1,0 +1,29 @@
+package edu.ntnu.idi.idattx2002.logic.Board.Actions;
+
+import edu.ntnu.idi.idattx2002.logic.Board.LandAction;
+import edu.ntnu.idi.idattx2002.logic.Player.Player;
+
+public class LadderAction implements LandAction {
+
+  private final int destinationTileId;
+
+  public LadderAction (int destinationTileId) {
+    validateDestinationTileId(destinationTileId);
+    this.destinationTileId = destinationTileId;
+  }
+
+  private void validateDestinationTileId(int destinationTileId) {
+    if (destinationTileId < 1) {
+      throw new IllegalArgumentException("destinationTile has to be greater than zero");
+    }
+  }
+
+  public int getDestinationTileId() {
+    return destinationTileId;
+  }
+
+  @Override
+  public void perform(Player player) {
+    player.movePlayerToTile(this.destinationTileId);
+  }
+}
