+++ src/main/java/edu/ntnu/idi/idattx2002/logic/Board/Actions/SwitchWithRandomAction.java
@@ -1,0 +1,33 @@
+package edu.ntnu.idi.idattx2002.logic.Board.Actions;
+
+import java.util.Random;
+
+import static edu.ntnu.idi.idattx2002.logic.Games.SnakesAndLadders.players;
+import edu.ntnu.idi.idattx2002.logic.Board.LandAction;
+import edu.ntnu.idi.idattx2002.logic.Player.Player;
+
+public class SwitchWithRandomAction implements LandAction {
+
+  private final Random random;
+
+  public SwitchWithRandomAction() {
+    random = new Random();
+  }
+
+  @Override
+  public void perform(Player player) {
+      Player playerToSwitch;
+      do {
+          playerToSwitch = players.get(random.nextInt(players.size()) + 1);
+      } while (playerToSwitch.getPlayerID() == player.getPlayerID());
+
+      int playerTileId = player.getCurrentTile();
+      int playerToSwitchTileId = playerToSwitch.getCurrentTile();
+
+      player.movePlayerToTile(playerToSwitchTileId);
+      playerToSwitch.movePlayerToTile(playerTileId);
+  }
+
+
+
+}
