+++ src/main/java/edu/ntnu/idi/idattx2002/logic/Dice/Dice.java
@@ -1,0 +1,39 @@
+package edu.ntnu.idi.idattx2002.logic.Dice;
+import edu.ntnu.idi.idattx2002.logic.DiceObserver;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
+public class Dice {
+
+  private final Random randomNumberGenerator;
+
+  private List<DiceObserver> observers;
+  private int currentValue;
+
+  public Dice() {
+    randomNumberGenerator = new Random();
+    observers = new ArrayList<>();
+    currentValue = 6;
+  }
+
+  public int getCurrentValue() {
+    return currentValue;
+  }
+
+  public int throwDice() {
+    currentValue = randomNumberGenerator.nextInt(6) + 1;
+    notifyObservers();
+    return currentValue;
+  }
+
+  public void addObserver(DiceObserver diceObserver) {
+    observers.add(diceObserver);
+  }
+
+  public void notifyObservers() {
+    for (DiceObserver diceObserver : observers) {
+      diceObserver.update(this);
+    }
+  }
+}
