+++ src/test/java/DiceTest.java
@@ -1,4 +1,4 @@
-import edu.ntnu.idi.idattx2002.Modules.Dice.Dice;
+import edu.ntnu.idi.idattx2002.logic.Dice.Dice;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
